% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/errors.R
\name{condition}
\alias{condition}
\title{Constructor for custom condition class}
\usage{
condition(subclass, message, call = sys.call(-1), ...)
}
\arguments{
\item{subclass}{the subclass for the condition}

\item{message}{character string.}

\item{call}{call expression.}

\item{...}{other classes}
}
\description{
Constructor for custom condition class.
}
\keyword{internal}
