% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dirichlet.R
\name{fun_rdirichlet}
\alias{fun_rdirichlet}
\title{Sample n times from a Dirichlet}
\usage{
fun_rdirichlet(n, a, text = "x")
}
\arguments{
\item{n}{number of samples}

\item{a}{Dirichlet parameter (a vector)}

\item{text}{optional variable names (default: 'x')}
}
\value{
a tibble with named colums, where each row is a Dirichlet sample
}
\description{
Sample n times from a Dirichlet
}
