% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normal.R
\name{fun_rnorm_population}
\alias{fun_rnorm_population}
\title{Generate a Normal-Normal population.}
\usage{
fun_rnorm_population(n, m, list_hyper = NULL, name_var = "x",
  name_source = list(mu = "mu", sigma = "sigma"))
}
\arguments{
\item{n}{number of samples per source}

\item{m}{number of sources}

\item{list_hyper}{a list containing the hyperparameters:
\itemize{
\item 'mu_mu0', 'mu_sigma0': the between-source mean hyperparameters
\item 'sigma_mu0', 'sigma_sigma0': the between-source sd hyperparameters
A partial list will be merged.
If \code{NULL}, they are generated from the N(0,1) distribution.
}}

\item{name_var}{names for data variables (default: \code{'x'})}

\item{name_source}{named list or character vector with names for source parameters \eqn{\mu} and \eqn{\sigma} (default: \code{list(mu = 'mu', sigma = 'sigma')})}
}
\value{
list of samples:
\itemize{
\item \code{list_hyper}: a list containing the hyperparameters:
\item \code{df_sources}: tibble of the Dirichlet population parameters, source column is \code{'source'}, variables start with \code{name_source}
\item \code{df_pop}: the Dirichlet data, source column is \code{'source'}, variables start with \code{name_var}
\item \code{names_var}: names of columns containing data variables
\item \code{names_source}: names of columns containing source variables
}
}
\description{
Generate samples from m sources and p parameters, n sample per source.
Optionally pass the between-source hyperparameters used to generate the source parameters.
}
\section{Model}{


The Normal-Normal model:

\deqn{X_{ij} ~ N(\mu_i, \sigma_i^2)  i = 1, \ldots, n, j = 1, \ldots, m}
\deqn{\mu_i ~ N(\mu^{\mu}_0, \sigma^{\mu}_0^2) j = 1, \ldots, m}
\deqn{\sigma_i ~ N(\mu^{\sigma}_0, \sigma^{\sigma}_0^2) j = 1, \ldots, m}
}

\seealso{
Other population functions: \code{\link{fun_rdirichlet_population}}
}
\concept{population}
\concept{population functions}
