% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general.R
\name{fun_var_names}
\alias{fun_var_names}
\title{Generate variable names}
\usage{
fun_var_names(p, text = "x")
}
\arguments{
\item{p}{number of variables}

\item{text}{variable name (default: 'x')}
}
\value{
variable names as strings
}
\description{
Generated variable names follow the rule "text[index]", with index from 1 to p.
}
