% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ref_quest_split.R
\name{make_dataset_splits}
\alias{make_dataset_splits}
\title{Generate reference/questioned/background observations from a multiple-source dataset.}
\usage{
make_dataset_splits(df, k_ref, k_quest, col_source = "source", ...)
}
\arguments{
\item{df}{all available data}

\item{k_ref}{number of reference samples}

\item{k_quest}{number of questioned samples}

\item{col_source}{column containing the source identifier (string or column number)}

\item{...}{Arguments passed on to \code{make_idx_splits}
\describe{
  \item{sources}{all class labels}
  \item{source_ref}{the reference source (scalar; if \code{NULL}, a random source will be picked)}
  \item{source_quest}{the questioned source(s) (if \code{NULL}, anything but the reference source: behaviour overridden by \code{same_source})}
  \item{same_source}{if \code{source_quest} is \code{NULL} and \code{same_source} is \code{TRUE}, questioned source is the reference source}
  \item{k_ref}{number of reference samples}
  \item{k_quest}{number of questioned samples}
  \item{background}{see details (default: \code{'outside'})}
  \item{replace}{use sampling with replacement, else error}
}}

\item{source_ref}{the reference source (scalar; if \code{NULL}, a random source will be picked)}

\item{source_quest}{the questioned source(s) (if \code{NULL}, anything but the reference source)}

\item{background}{see details (default: \code{outside})}

\item{replace}{use sampling with replacement, else error}
}
\value{
a list of indexes (\code{idx_reference}, \code{idx_questioned}, \code{idx_background}) and a list of dataframes (\code{df_reference}, \code{df_questioned}, \code{df_background})
}
\description{
The function splits a dataframe by rows, into a sample of reference items, questioned items and background items.
The split is done according by the item source.
}
\details{
Reference and questioned samples are always non-intersecting, even when the source is the same.

Sampling with replacement is used, if necessary and not forbidden.
}
\section{Background selection}{

\itemize{
\item If \code{background} is \code{'outside'}, the background dataset comprises all items who \strong{do not lie} in any of the reference and questioned sets.
\item If \code{background} is \code{'others'}, the background dataset comprises all items from the \strong{non}-reference and \strong{non}-questioned \strong{sources}
}

By default, \code{background} is \code{'outside'}.
}

\section{Source sampling}{


If \code{source_quest} is \code{NULL}:
\itemize{
\item if \code{same_source} is \code{NULL} or \code{FALSE}, questioned items are sampled from all but the reference source.
\item if \code{same_source} is \code{TRUE}, questioned items are sampled from the reference source.
}

Else, questioned items will be sampled from the questioned source(s), even if it contains the reference one.

Items will never be sampled once (unless \code{replace} is \code{TRUE}): they appear once in the reference/questioned/background items.
}

\seealso{
\code{\link[=make_idx_splits]{make_idx_splits()}}
}
