% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ref_quest_split.R
\name{resample}
\alias{resample}
\title{Random Samples and Permutations}
\usage{
resample(x, size, replace = FALSE)
}
\arguments{
\item{x}{a vector of one or more elements from which to choose}

\item{size}{a non-negative integer giving the number of items to choose.}

\item{replace}{should sampling be with replacement?}
}
\description{
\code{resample} takes a sample of the specified size from the elements of \code{x} using either with or without replacement.
}
\details{
Works like \code{\link[base:sample]{base::sample()}}, but it is safer if \code{x} is a scalar.
}
