% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/errors.R
\name{stop_fatal}
\alias{stop_fatal}
\title{Raise a FATAL error}
\usage{
stop_fatal(message, call = sys.call(-1), ...)
}
\arguments{
\item{message}{character string.}

\item{call}{call expression.}

\item{...}{other classes}
}
\description{
Raise a FATAL error. Useful in testing.
These errors should never happen.
}
\keyword{internal}
