% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/errors.R
\name{stop_serious}
\alias{stop_serious}
\title{Raise a SERIOUS error}
\usage{
stop_serious(message, call = sys.call(-1), ...)
}
\arguments{
\item{message}{character string.}

\item{call}{call expression.}

\item{...}{other classes}
}
\description{
Raise a SERIOUS error.
These errors might be expected.
}
\keyword{internal}
