% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dirichlet.R
\name{fun_rdirichlet_population}
\alias{fun_rdirichlet_population}
\title{Generate samples from m sources and p parameters, n sample per source.}
\usage{
fun_rdirichlet_population(n, m, p, alpha = NULL, name_var = "x",
  name_source = "theta")
}
\arguments{
\item{n}{number of samples per source}

\item{m}{number of sources}

\item{p}{number of variables}

\item{alpha}{between-source alpha hyperparameter.
If NULL, it is generated from the Uniform distribution on the (p-1)-simplex.}

\item{name_var}{names for data variables (default: 'x')}

\item{name_source}{names for source parameters (default: 'theta')}
}
\value{
list of samples:
- alpha: the Dirichlet between hyperparameter
- df_sources: tibble of the Dirichlet population parameters, source column is 'source', variables start with name_source
- df_pop: the Dirichlet data, source column is 'source', variables start with name_var
- names_var: names of columns containing data variables
- names_source: names of columns containing source variables
}
\description{
Optionally pass the between-source alpha hyperparameter used to generate the source parameters.
}
