% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ref_quest_split.R
\name{make_dataset_splits}
\alias{make_dataset_splits}
\title{Generate reference and questioned observations from}
\usage{
make_dataset_splits(df, k_ref, k_quest, col_source = "source", ...)
}
\arguments{
\item{df}{all available data}

\item{k_ref}{number of reference samples}

\item{k_quest}{number of questioned samples}

\item{col_source}{column containing the source identifier}
}
\value{
a list of indexes and a list of dataframes
}
\description{
Generate reference and questioned observations from
}
