% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ref_quest_split.R
\name{make_dataset_splits}
\alias{make_dataset_splits}
\title{Generate reference/questioned/background observations from a multiple-source dataset.}
\usage{
make_dataset_splits(df, k_ref, k_quest, col_source = "source", ...)
}
\arguments{
\item{df}{all available data}

\item{k_ref}{number of reference samples}

\item{k_quest}{number of questioned samples}

\item{col_source}{column containing the source identifier (string or column number)}

\item{...}{Arguments passed on to \code{make_idx_splits}
\describe{
  \item{source_ref}{the reference source (scalar; if \code{NULL}, a random source will be picked)}
  \item{source_quest}{the questioned source(s) (if \code{NULL}, anything but the reference source: behaviour overridden by \code{same_source})}
  \item{same_source}{if \code{source_quest} is \code{NULL} and \code{same_source} is \code{TRUE}, questioned source is the reference source}
  \item{source_ref_allowed}{allowed reference sources (if not \code{NULL} (default), reference source will be picked among these)}
  \item{source_quest_allowed}{allowed  questioned sources (if not \code{NULL} (default), questioned source(s) will be picked among these)}
  \item{background}{see details (default: \code{'outside'})}
  \item{replace}{use sampling with replacement, else error}
}}
}
\value{
a list of indexes (\code{idx_reference}, \code{idx_questioned}, \code{idx_background}) and a list of dataframes (\code{df_reference}, \code{df_questioned}, \code{df_background})
}
\description{
The function splits a dataframe by rows, into a sample of reference items, questioned items and background items.
The split is done according by the item source.
}
\details{
Reference and questioned samples are always non-intersecting, even when the source is the same.

Sampling with replacement is used, if necessary and not forbidden.
}
\section{Background selection}{

\itemize{
\item If \code{background} is \code{'outside'}, the background dataset comprises all items who \strong{do not lie} in any of the reference and questioned sets.
It can contain items from reference and questioned sources.
\item If \code{background} is \code{'others'}, the background dataset comprises all items from the \strong{non}-reference and \strong{non}-questioned \strong{potential sources}.
\item If \code{background} is \code{'unobserved'}, the background dataset comprises all items from the sources who do not appear in any of the reference and questioned sets.
}

By default, \code{background} is \code{'outside'}.

Notice that \code{background = 'others'} generates no background data if questioned sources are not specified:
the union of reference and questioned sources fills the available sources in the population.
}

\section{Source sampling}{


By design, the package identifies sources which are allowed to be sampled from.
By default, all available sources can appear in the reference or questioned samples.

This restriction can be modified using the parameters \code{source_ref_allowed} and \code{source_quest_allowed}.

The behaviour of \code{same_source} is stronger, and \code{source_quest_allowed} is overridden.

If \code{source_quest} is \code{NULL}:
\itemize{
\item if \code{same_source} is \code{NULL} or \code{FALSE}, questioned items are sampled from all but the reference source (or \code{source_quest_allowed}).
\item if \code{same_source} is \code{TRUE}, questioned items are sampled from the reference source (\code{source_ref_allowed} is ignored).
}

Else, questioned items will be sampled from the questioned source(s), even if it contains the reference one.

It is also forbidden to specify a \code{source_ref} or a \code{source_quest} which are not allowed.

Items will never be sampled once (unless \code{replace} is \code{TRUE}): they appear once in the reference/questioned/background items.
}

\examples{
\dontrun{
# Sample different species
make_dataset_splits(iris, 5, 5, col_source = 'Species')

# Sample same species
make_dataset_splits(iris, 5, 5, col_source = 'Species', same_source = TRUE)

# Sample from custom species
make_dataset_splits(iris, 5, 5, col_source = 'Species',
   source_ref = 'virginica', source_quest = 'versicolor')
make_dataset_splits(iris, 5, 5, col_source = 'Species',
   source_ref = 'virginica', source_quest = c('virginica', 'versicolor'))

# Sample from reference source with replacement
make_dataset_splits(iris, 500, 5, col_source = 'Species', replace = TRUE)

# Use background sources from non-sampled items
make_dataset_splits(iris, 50, 50, col_source = 'Species',
   source_ref = 'virginica', source_quest = 'versicolor', background = 'others')
}
}
\seealso{
Other set sampling functions: \code{\link{make_idx_splits}}
}
\concept{set sampling}
\concept{set sampling functions}
