% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ref_quest_split.R
\name{make_idx_splits}
\alias{make_idx_splits}
\title{Extract questioned/reference/background samples from a list of populations from sources}
\usage{
make_idx_splits(sources, k_ref, k_quest, source_ref = NULL,
  source_quest = NULL, same_source = NULL, background = "outside",
  replace = TRUE)
}
\arguments{
\item{sources}{all class labels}

\item{k_ref}{number of reference samples}

\item{k_quest}{number of questioned samples}

\item{source_ref}{the reference source (scalar; if \code{NULL}, a random source will be picked)}

\item{source_quest}{the questioned source(s) (if \code{NULL}, anything but the reference source: behaviour overridden by \code{same_source})}

\item{same_source}{if \code{source_quest} is \code{NULL} and \code{same_source} is \code{TRUE}, questioned source is the reference source}

\item{background}{see details (default: \code{'outside'})}

\item{replace}{use sampling with replacement, else error}
}
\value{
list of indexes (\code{idx_reference}, \code{idx_questioned}, \code{idx_background})
}
\description{
The function splits a list of items (rows) into a sample of reference items, questioned items and background items.
}
\details{
Reference/questioned/background samples are always non-intersecting, even when the source is the same.

Sampling with replacement is used, if necessary and not forbidden. If it is used, a message appears.

It is \strong{always} guaranteed that no sample appear more than once across reference/questioned/background items (but it can appear multiple times in a set).
}
\section{Source sampling}{


If \code{source_quest} is \code{NULL}:
\itemize{
\item if \code{same_source} is \code{NULL} or \code{FALSE}, questioned items are sampled from all but the reference source.
\item if \code{same_source} is \code{TRUE}, questioned items are sampled from the reference source.
}

Else, questioned items will be sampled from the questioned source(s), even if it contains the reference one.

Items will never be sampled once (unless \code{replace} is \code{TRUE}): they appear once in the reference/questioned/background items.
}

\section{Background selection}{

\itemize{
\item If \code{background} is \code{'outside'}, the background dataset comprises all items who \strong{do not lie} in any of the reference and questioned sets.
It can contain items from reference and questioned sources.
\item If \code{background} is \code{'others'}, the background dataset comprises all items from the \strong{non}-reference and \strong{non}-questioned \strong{potential sources}.
\item If \code{background} is \code{'unobserved'}, the background dataset comprises all items from the sources who do not appear in any of the he reference and questioned sets.
}

By default, \code{background} is \code{'outside'}.

Notice that \code{background = 'others'} generates no background data if questioned sources are not specified:
the union of reference and questioned sources fills the available sources in the population.
}

\seealso{
Other set sampling functions: \code{\link{make_dataset_splits}}
}
\concept{set sampling}
\concept{set sampling functions}
